+++ sensor.h
@@ -3,27 +3,27 @@
 
 #include <cstdlib>
 #include <iostream>
-#include <string>
+//#include <string>
 #include <array.h>
 
 using namespace std;
 
 class Sensor {
 	string ID;
-	Array<float> data;
+	Array<double> data;
 public:
 	Sensor();
 	Sensor(const string&);
 	Sensor(const Sensor&);
 	~Sensor();
 	
-	float getData(int);
+	double getData(int);
 	int getSize() const;
 	const string& getID()const;
 	void clear();
 	void querry(ostream&,int,int);
 	
-	Sensor& operator+(const float);
+	Sensor& operator+(const double);
 	Sensor& operator+(const Sensor&);
 	Sensor& operator=(const string&);
 	Sensor& operator=(const Sensor&);
